refactor: move controllers and services to server
