+++ src/app/api/steps/route.js
@@ -1,11 +1,9 @@
 import { NextResponse } from 'next/server'
-import { getUserByRoleTeacherServiceJson } from '@/app/services/userService'
-import { getByUserIdInMembers } from '@/app/services/tccService'
-import { getTccInfoByUserIdInMembers } from '@/app/controllers/tcc/readTccActions'
+import { getTccInfoByUserIdInMembersAction } from '@/server/controllers/tcc/readTccActions'
 
 export async function GET() {
     try {
-        const tcc = await getTccInfoByUserIdInMembers()
+        const tcc = await getTccInfoByUserIdInMembersAction()
         return tcc
     } catch (error) {
         return NextResponse.json({ data: [] }, { status: 500 })
