+++ src/server/controllers/theme/createThemeActions.js
@@ -1,20 +1,18 @@
 'use server'
 import { z } from 'zod'
-import { getThemeByIdService, updateThemeByIdService } from '@/app/services/themeService'
+import { createThemeService } from '@/server/services/themeService'
 import { themeSchema } from '@/utils/schemas/themeSchema'
 
-export async function updateThemeActions(id, formData) {
+export async function createThemeActions(formData) {
     try {
-        const { data: theme } = await getThemeByIdService(id)
-
         const formDataObj = {
-            name: formData.get('name') || theme.name,
-            description: formData.get('description') || theme.description,
+            name: formData.get('name'),
+            description: formData.get('description'),
         }
 
         const validatedData = themeSchema.parse(formDataObj)
 
-        const result = await updateThemeByIdService(id, validatedData)
+        const result = await createThemeService(validatedData)
         return result
     } catch (error) {
         if (error instanceof z.ZodError) {
@@ -25,7 +23,7 @@ export async function updateThemeActions(id, formData) {
         if (error instanceof TypeError) {
             return { success: false, message: 'Erro de tipo de dado. Verifique os dados fornecidos.' }
         } else {
-            return { success: false, message: error.message }
+            return { success: false, message: 'Erro inesperado ao criar usuário.' }
         }
     }
 }
