+++ src/server/controllers/theme/updateThemeActions.js
@@ -1,18 +1,20 @@
 'use server'
 import { z } from 'zod'
-import { createThemeService } from '@/app/services/themeService'
+import { getThemeByIdService, updateThemeByIdService } from '@/server/services/themeService'
 import { themeSchema } from '@/utils/schemas/themeSchema'
 
-export async function createThemeActions(formData) {
+export async function updateThemeActions(id, formData) {
     try {
+        const { data: theme } = await getThemeByIdService(id)
+
         const formDataObj = {
-            name: formData.get('name'),
-            description: formData.get('description'),
+            name: formData.get('name') || theme.name,
+            description: formData.get('description') || theme.description,
         }
 
         const validatedData = themeSchema.parse(formDataObj)
 
-        const result = await createThemeService(validatedData)
+        const result = await updateThemeByIdService(id, validatedData)
         return result
     } catch (error) {
         if (error instanceof z.ZodError) {
@@ -23,7 +25,7 @@ export async function createThemeActions(formData) {
         if (error instanceof TypeError) {
             return { success: false, message: 'Erro de tipo de dado. Verifique os dados fornecidos.' }
         } else {
-            return { success: false, message: 'Erro inesperado ao criar usuário.' }
+            return { success: false, message: error.message }
         }
     }
 }
