+++ src/server/services/libraryService.js
